Remove nose from validation_efield_mpi, validation_rd_mpi and validation_rd_dist
