+++ validation/run_validation_mpi_tests.py
@@ -1,32 +1,27 @@
-from __future__ import absolute_import
-
+import contextlib
 import os.path as osp
 import sys
-import collections.abc
-collections.Callable = collections.abc.Callable
-
-import nose
+import unittest
 
 from config import configuration
 
-
-
-
-DEFAULT_TEST_SUITE = [
-    'validation_rd_mpi/unbdiff2D.py',
-    'validation_rd_mpi/unbdiff2D_linesource_ring.py',
-    'validation_rd_mpi/unbdiff.py',
-    'validation_rd_mpi/bounddiff.py',
-    'validation_rd_mpi/csd_clamp.py',
-    'validation_rd_mpi/masteq_diff.py',
-    'validation_rd_mpi/kisilevich.py',
-    'validation_efield_mpi/test_rallpack1_dist.py',
-]
-
-
 if __name__ == '__main__':
-    test_dir = osp.dirname(osp.abspath(__file__))
-    test_suite = sys.argv[1:] or DEFAULT_TEST_SUITE
     configuration.suffix = '_mpi'
-    for suite in test_suite:
-        nose.run(argv=['-s', '-v', osp.join(test_dir, suite)])
+    test_dir = osp.dirname(osp.abspath(__file__))
+    loader = unittest.TestLoader()
+
+    directories = []
+    for fold in sys.argv[1:]:
+        directories.append(osp.join(test_dir, fold))
+    if len(directories) == 0:
+        directories.append(test_dir)
+
+    for start_dir in directories:
+        print(f'Running validations from {start_dir}')
+
+        with contextlib.redirect_stdout(None):
+            top_dir = osp.join(test_dir, '..')
+            suite = loader.discover(test_dir, pattern='parallel_*.py', top_level_dir=top_dir)
+
+        runner = unittest.TextTestRunner(verbosity=2)
+        runner.run(suite)
